chore: move GCP detector into detectors file
