+++ packages/opentelemetry-node/lib/detectors.js
@@ -7,7 +7,9 @@
  * @typedef {import('@opentelemetry/resources').ResourceDetector} ResourceDetector
  */
 
-const {getStringListFromEnv, suppressTracing} = require('@opentelemetry/core');
+const {URL} = require('url');
+
+const { getStringListFromEnv, suppressTracing } = require('@opentelemetry/core');
 const { context } = require('@opentelemetry/api');
 
 const {
@@ -47,9 +49,9 @@ const {
     SEMRESATTRS_K8S_NAMESPACE_NAME,
     SEMRESATTRS_K8S_POD_NAME,
 } = require('@opentelemetry/semantic-conventions');
+const jsonBigint = require('json-bigint');
 
-const {log} = require('./logging');
-const { gcpDetector } = require('./detector-gcp');
+const { log } = require('./logging');
 
 // @ts-ignore - compiler options do not allow lookp outside `lib` folder
 const ELASTIC_SDK_VERSION = require('../package.json').version;
@@ -69,6 +71,52 @@ const distroDetector = {
         };
     },
 };
+
+// TODO: Switch back to `@opentelemetry/resource-detector-gcp` when
+// https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2320 is complete
+/** @type {ResourceDetector} */
+const gcpDetector = {
+    detect() {
+        const gcpMetadataBaseUrl = new URL('/', 'http://metadata.google.internal:80');
+        const metadataUrl = gcpMetadataBaseUrl + 'computeMetadata/v1/?recursive=true';
+        const options = {
+            method: 'GET',
+            headers: { 'Metadata-Flavor': 'Google' },
+            signal: AbortSignal.timeout(1000),
+        };
+
+        const metadataPromise = context.with(suppressTracing(context.active()), () =>
+            fetch(metadataUrl, options)
+            .then(res => res.text())
+            .then(txt => jsonBigint.parse(txt))
+            .catch((err) => {
+                log.debug({ err }, 'Unable to get GCP metadata');
+                return undefined;
+            })
+        );
+
+        // TODO: switch to `@opentelemetry/resource-detector-gcp` when the below issue is fixed
+        // https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2320
+        const attributes = {
+            [SEMRESATTRS_CLOUD_PROVIDER]: metadataPromise.then(md => md && CLOUDPROVIDERVALUES_GCP),
+            [SEMRESATTRS_CLOUD_ACCOUNT_ID]: metadataPromise.then(md => md?.project?.projectId),
+            [SEMRESATTRS_HOST_ID]: metadataPromise.then(md => md?.instance?.id),
+            [SEMRESATTRS_HOST_NAME]: metadataPromise.then(md => md?.instance?.hostname),
+            [SEMRESATTRS_CLOUD_AVAILABILITY_ZONE]: metadataPromise.then(md => md?.zone),
+        };
+
+        // Add resource attributes for K8s.
+        if (process.env.KUBERNETES_SERVICE_HOST) {
+            // attributes[SEMRESATTRS_K8S_CLUSTER_NAME] = this._getClusterName(isAvail); MISSING
+            attributes[SEMRESATTRS_K8S_NAMESPACE_NAME] = metadataPromise.then(md => md && process.env.NAMESPACE);
+            attributes[SEMRESATTRS_K8S_POD_NAME] = metadataPromise.then(md => md && process.env.HOSTNAME);
+            attributes[SEMRESATTRS_CONTAINER_NAME] = metadataPromise.then(md => md && process.env.CONTAINER_NAME);
+        }
+        return { attributes };
+    },
+}
+
+
 
 /** @type {Record<string, ResourceDetector | Array<ResourceDetector>>} */
 const defaultDetectors = {
